better server side organization
