+++ src/server/routing/v1/websocket/filetree.js
@@ -1,0 +1,77 @@
+import http from "node:http";
+import { WebSocketServer } from "ws";
+import { FILETREE_PREFIX, OTHandler } from "./ot-handler.js";
+
+/**
+ * Set up file-tree related websocket handling given
+ * an express app, and the session parser (which we
+ * very much need to make sure only authenticated
+ * users are allowed to establish a socket connection)
+ */
+export function setupFileTreeWebSocket(app, sessionParser) {
+  const server = http.createServer(app);
+  const wss = new WebSocketServer({ clientTracking: false, noServer: true });
+  server.on("upgrade", (request, socket, head) => {
+    sessionParser(request, {}, () => {
+      const { user } = request.session.passport ?? {};
+      if (!user) {
+        socket.write(`HTTP/1.1 401 Unauthorized\r\n\r\n`);
+        socket.destroy();
+        return;
+      }
+      console.log(`Session is parsed, user is ${user.id}!`);
+      wss.handleUpgrade(request, socket, head, (ws) => {
+        wss.emit(`connection`, ws, request);
+      });
+    });
+  });
+
+  wss.on("connection", (socket, request) => {
+    addFileTreeCommunication(socket, request);
+  });
+
+  return server;
+}
+
+/**
+ * Anyone can use this function to tack file-tree compatible
+ * message handling to a websocket.
+ */
+export async function addFileTreeCommunication(socket, request) {
+  // you are you *sure* you're allowed in here?
+  if (!request.session?.passport?.user) return;
+
+  // Our websocket based request handler.
+  const otHandler = new OTHandler(socket, request.session.passport.user);
+
+  socket.on("message", async (message) => {
+    // This will not throw, because a server shouldn't crash out.
+    let data = message.toString();
+    try {
+      data = JSON.parse(data);
+    } catch (e) {
+      console.warn(
+        `Received incompatible data via websocket: message is not JSON.`,
+        data,
+      );
+    }
+    if (!data) return;
+
+    // Is this something we know how to handle?
+    let { type, detail } = data;
+    if (!type) return;
+    if (!type.startsWith(FILETREE_PREFIX)) return;
+
+    // Looks like it, let's get it processed.
+    type = type.replace(FILETREE_PREFIX, ``);
+    const handlerName = `on${type}`;
+    const handler = otHandler[handlerName].bind(otHandler);
+    if (!handler) {
+      return console.warn(`Missing implementation for ${handlerName}.`);
+    }
+
+    // TODO: what do we do if the container's been put to sleep??
+
+    handler(detail, request);
+  });
+}
